requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
pyserial>=3.5
imutils>=0.5.4
